refactor: Moved all burst services to the root services directory
