+++ src/BurstChat.Client/src/app/components/core/incoming-call/incoming-call.component.ts
@@ -5,8 +5,8 @@ import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { faCheck, faTimes, faUserCircle } from '@fortawesome/free-solid-svg-icons';
 import { User } from 'src/app/models/user/user';
 import { RTCSessionContainer } from 'src/app/models/chat/rtc-session-container';
-import { UserService } from 'src/app/modules/burst/services/user/user.service';
-import { RtcSessionService } from 'src/app/modules/burst/services/rtc-session/rtc-session.service';
+import { UserService } from 'src/app/services/user/user.service';
+import { RtcSessionService } from 'src/app/services/rtc-session/rtc-session.service';
 import { CardComponent } from 'src/app/components/shared/card/card.component';
 import { CardHeaderComponent } from 'src/app/components/shared/card-header/card-header.component';
 import { CardBodyComponent } from 'src/app/components/shared/card-body/card-body.component';
