+++ src/BurstChat.Client/src/app/components/core/layout/layout.component.ts
@@ -3,14 +3,14 @@ import { RouterOutlet } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { Invitation } from 'src/app/models/servers/invitation';
 import { NotifyService } from 'src/app/services/notify/notify.service';
-import { SidebarService } from 'src/app/modules/burst/services/sidebar/sidebar.service';
-import { UserService } from 'src/app/modules/burst/services/user/user.service';
-import { ServersService } from 'src/app/modules/burst/services/servers/servers.service';
-import { ChannelsService } from 'src/app/modules/burst/services/channels/channels.service';
-import { DirectMessagingService } from 'src/app/modules/burst/services/direct-messaging/direct-messaging.service';
-import { ChatService } from 'src/app/modules/burst/services/chat/chat.service';
-import { MediaService } from 'src/app/modules/burst/services/media/media.service';
-import { RtcSessionService } from 'src/app/modules/burst/services/rtc-session/rtc-session.service';
+import { SidebarService } from 'src/app/services/sidebar/sidebar.service';
+import { UserService } from 'src/app/services/user/user.service';
+import { ServersService } from 'src/app/services/servers/servers.service';
+import { ChannelsService } from 'src/app/services/channels/channels.service';
+import { DirectMessagingService } from 'src/app/services/direct-messaging/direct-messaging.service';
+import { ChatService } from 'src/app/services/chat/chat.service';
+import { MediaService } from 'src/app/services/media/media.service';
+import { RtcSessionService } from 'src/app/services/rtc-session/rtc-session.service';
 import { SidebarComponent } from 'src/app/components/core/sidebar/sidebar.component';
 import { IncomingCallComponent } from 'src/app/components/core/incoming-call/incoming-call.component';
 import { AddServerComponent } from 'src/app/components/core/add-server/add-server.component';
