+++ src/BurstChat.Client/src/app/modules/chat/components/chat-info/chat-info.component.ts
@@ -6,8 +6,8 @@ import { PrivateGroupConnectionOptions } from 'src/app/models/chat/private-group
 import { DirectMessagingConnectionOptions } from 'src/app/models/chat/direct-messaging-connection-options';
 import { User } from 'src/app/models/user/user';
 import { RTCSessionContainer } from 'src/app/models/chat/rtc-session-container';
-import { UserService } from 'src/app/modules/burst/services/user/user.service';
-import { RtcSessionService } from 'src/app/modules/burst/services/rtc-session/rtc-session.service';
+import { UserService } from 'src/app/services/user/user.service';
+import { RtcSessionService } from 'src/app/services/rtc-session/rtc-session.service';
 import { UiLayerService } from 'src/app/modules/chat/services/ui-layer/ui-layer.service';
 
 import {
