+++ src/BurstChat.Client/src/app/modules/servers/components/edit-server-users/edit-server-users.component.ts
@@ -3,9 +3,9 @@ import { Subscription } from 'rxjs';
 import { Server } from 'src/app/models/servers/server';
 import { User } from 'src/app/models/user/user';
 import { NotifyService } from 'src/app/services/notify/notify.service';
-import { ServersService } from 'src/app/modules/burst/services/servers/servers.service';
-import { ChatService } from 'src/app/modules/burst/services/chat/chat.service';
-import { UserService } from 'src/app/modules/burst/services/user/user.service';
+import { ServersService } from 'src/app/services/servers/servers.service';
+import { ChatService } from 'src/app/services/chat/chat.service';
+import { UserService } from 'src/app/services/user/user.service';
 
 /**
  * This class represents an angular component that enables invitings and editing server users.
