+++ src/BurstChat.Client/src/app/modules/servers/components/edit-server/edit-server.component.ts
@@ -3,8 +3,8 @@ import { ActivatedRoute } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { Server } from 'src/app/models/servers/server';
 import { DisplayServer } from 'src/app/models/sidebar/display-server';
-import { ServersService } from 'src/app/modules/burst/services/servers/servers.service';
-import { SidebarService } from 'src/app/modules/burst/services/sidebar/sidebar.service';
+import { ServersService } from 'src/app/services/servers/servers.service';
+import { SidebarService } from 'src/app/services/sidebar/sidebar.service';
 
 /**
  * This class represents an angular component that presents information about the active server and enables
