+++ src/BurstChat.Client/src/app/services/sidebar/sidebar.service.ts
